id added to models

+++ Controllers/LikesController.cs
@@ -8,6 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using instaclone.Data;
 using instaclone.models;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace instaclone.Controllers
 {
@@ -24,62 +25,54 @@ namespace instaclone.Controllers
         }
 
         // GET: api/Likes
+        // get all likes from current user
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Like>>> GetLikes()
         {
-            return await _context.Likes.ToListAsync();
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userId == null)
+                return BadRequest();
+
+            return await _context.Likes.Where(l => l.InstaCloneUser.Id == userId.Value).ToListAsync();
         }
+
 
         // GET: api/Likes/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Like>> GetLike(int id)
+        //[HttpGet("{id}")]
+        //public async Task<ActionResult<Like>> GetLike(int id)
+        //{
+        //    var like = await _context.Likes.FindAsync(id);
+
+        //    if (like == null)
+        //    {
+        //        return NotFound();
+        //    }
+
+        //    return like;
+        //}
+
+        // GET: api/Likes/Post/5
+        // Get all likes for a post
+        [HttpGet("Post/{id}")]
+        public async Task<ActionResult<IEnumerable<Like>>> GetLikesForPost(int id)
         {
-            var like = await _context.Likes.FindAsync(id);
-
-            if (like == null)
-            {
-                return NotFound();
-            }
-
-            return like;
+            return await _context.Likes.Where(l => l.Post.Id == id).ToListAsync();
         }
 
-        // PUT: api/Likes/5
+        // POST: api/Likes/Post/5
+        // add a like to a post
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutLike(int id, Like like)
+        [HttpPost("Post/{postId}")]
+        public async Task<ActionResult<Like>> PostLike(int postId)
         {
-            if (id != like.Id)
-            {
+            var user = await _context.InstaCloneUser.FindAsync(User.FindFirstValue(ClaimTypes.NameIdentifier));
+            var post = await _context.Posts.FindAsync(postId);
+
+            if (user == null || post == null)
                 return BadRequest();
-            }
 
-            _context.Entry(like).State = EntityState.Modified;
+            var like = new Like {InstaCloneUser = user, Post = post };
 
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!LikeExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return NoContent();
-        }
-
-        // POST: api/Likes
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<Like>> PostLike(Like like)
-        {
             _context.Likes.Add(like);
             await _context.SaveChangesAsync();
 
@@ -90,11 +83,15 @@ namespace instaclone.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteLike(int id)
         {
+            var user = await _context.InstaCloneUser.FindAsync(User.FindFirstValue(ClaimTypes.NameIdentifier));
             var like = await _context.Likes.FindAsync(id);
-            if (like == null)
-            {
-                return NotFound();
-            }
+
+            if (like == null || user == null)
+                return BadRequest();
+
+            if (!like.InstaCloneUser.Id.Equals(user.Id))
+                return Unauthorized();
+
 
             _context.Likes.Remove(like);
             await _context.SaveChangesAsync();
